+++ src/components/IconButton/IconButton.tsx
@@ -12,15 +12,31 @@ export interface IconButtonProps extends Omit<React.HTMLAttributes<HTMLButtonEle
    * How large should the button be?
    */
   size?: 'sm' | 'md' | 'lg';
+  /**
+   * Mainly used for styling
+   */
+  variant?: ButtonVariants;
+  /**
+   * Should the button be disabled?
+   */
+  isDisabled?: boolean;
 }
 
-const IconButton = ({ children, size = 'md', className, ...props }: IconButtonProps) => {
+const IconButton = ({
+  children,
+  size = 'md',
+  variant = ButtonVariants.primary,
+  isDisabled = false,
+  className,
+  ...props
+}: IconButtonProps) => {
   const { className: baseClass, ...commonProps } = getCommonProps(props, 'IconButton');
   return (
     <Button
       {...commonProps}
-      variant={ButtonVariants.primary}
-      className={classnames(baseClass, `${baseClass}--${size}`, className)}
+      variant={variant}
+      className={classnames(baseClass, `${baseClass}--${size}`, `${baseClass}--${variant}`, className)}
+      isDisabled={isDisabled}
       {...props}
     >
       {children}