+++ src/site-furniture/Header/Header.tsx
@@ -129,7 +129,11 @@ const Header = forwardRef<HTMLElement, HeaderProps>(
             {otherChildren}
           </HeaderContext.Provider>
         </div>
-        <div className={classnames(`${px}-header__overlay`, { [`${px}-header__overlay--active`]: isSearchExpanded })} />
+        <div
+          className={classnames(`${px}-header__overlay`, {
+            [`${px}-header__overlay--active`]: isSearchExpanded,
+          })}
+        />
       </header>
     );
   },